requirements_carla.txt
# First install these specific versions
# python -m pip install "typing-extensions==4.7.1" "numpy==1.19.4"
# python -m pip install torch===1.7.1 torchvision===0.8.2 torchaudio===0.7.2 -f https://download.pytorch.org/whl/torch_stable.html

# Install this 
protobuf<=3.21
lime==0.2.0.1
mip==1.12.0
pandas==1.1.4
recourse==1.0.0
scikit-learn==0.23.2
tensorflow==1.14.0
h5py==2.10.0
dice-ml==0.5
ipython
keras==2.3.0
xgboost==1.4.2
causalgraphicalmodels==0.0.4
carla-recourse